# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AAELink HTTP API",
    "version": "0.0.41-alpha",
    "summary": "Auto-generated route inventory. Method/path surface only — request/response schemas not yet introspected."
  },
  "servers": [
    {
      "url": "https://example.com",
      "description": "Replace with your AAELink deployment URL"
    }
  ],
...
  "paths": {
...
    "/api/approvals/requests": {
      "get": {
        "summary": "GET /api/approvals/requests",
        "tags": [
          "approvals"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "post": {
        "summary": "POST /api/approvals/requests",
        "tags": [
          "approvals"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/auth/entra": {
      "get": {
        "summary": "GET /api/auth/entra",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "summary": "GET /api/auth/me",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "PUT /api/auth/me",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/bootstrap": {
      "get": {
        "summary": "GET /api/bootstrap",
        "tags": [
          "bootstrap"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized — no session cookie"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  }
}